my_app/requirements.txt
async-generator==1.10
attrs==23.1.0
blinker==1.6.2
boto3==1.26.142
botocore==1.29.142
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docker==6.1.3
exceptiongroup==1.1.1
Flask==2.3.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
multi-key-dict==2.0.3
outcome==1.2.0
packaging==23.1
pbr==5.11.1
pluggy==1.0.0
py==1.11.0
PySocks==1.7.1
pytest==7.3.1
pytest-html==3.2.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jenkins==1.8.0
requests==2.31.0
s3transfer==0.6.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.16
webdriver-manager==3.8.6
websocket-client==1.5.2
Werkzeug==2.3.4
wsproto==1.2.0
